Rewrite adjacent-to, descendant-of, and child-of in terms of a generalized walking fn.
This new function hold the functionality necessary for all of these and probably
future selectors as well. It isn't as simple as I'd hoped, but it does the
trick. Also, added more complex tests to test the functionality of this, because
the complexity of the select-walk makes it fairly easy to break this function
with a simple change and not notice it without complex tests. These tests have
a standing TODO to improve the zip/next chains. It's pretty bad.