Unify transform pairs with a new concept: pews
This is a huge breaking change. First of all, we're getting rid of
separate selectors and transformers. They still appear to be separate,
but under the hood they get combined into one function that both
selects and transforms. In turn, we're also forcing users to pass in
vectors of selectors and transformer pairs *or* compose them themselves
into a pew.

This makes things much more consistent, simplifies the code a bit, and
makes it easier to add some upcoming features.