Use keyword opts for parsing fns and add :resource.
Because we have multiple options now, and since these functions are
user-facing and thus should have good APIs, I'm switching to keyword
params for giving them options. Will probably always just be :parser and
:resource. :resource is a convenience keyword for specifying that the
string given should be treated as a resource path. The reason I've done
this is because there is no other way to determine from just a string if
it should be treated as a resource or just raw HTML, and I'd like to not
have to force users to use the resource function directly.

I am aware that some people don't like keyword options very much, and
instead prefer numerous functions instead of options. I think these
options are simple enough and specific enough that it makes no sense to
clutter the namespace with confusing functions and make users figure out
which ones to use.